Refactor the API folder so its in the SRC folder
