+++ BackEnd/src/api/championstats.ts
@@ -1,9 +1,9 @@
 import express, { Response, Router } from 'express'
-import { TypedRequest } from '../../Common/Interface/Internal/responseUtil'
-import { CreatePlayerRadarWithName } from '../src/business/charts'
-import logger from '../logger'
-import { ChampionStatsSheetRequest } from '../../Common/Interface/Internal/championstats'
-import { GetChampionStatsBySeason } from '../src/business/championstats'
+import { TypedRequest } from '../../../Common/Interface/Internal/responseUtil'
+import { CreatePlayerRadarWithName } from '../business/charts'
+import logger from '../../logger'
+import { ChampionStatsSheetRequest } from '../../../Common/Interface/Internal/championstats'
+import { GetChampionStatsBySeason } from '../business/championstats'
 
 const router: Router = express.Router()
