+++ BackEnd/src/api/games.ts
@@ -1,11 +1,11 @@
 import express, { Router, Request } from 'express'
-import { TypedRequest, TypedResponse } from '../../Common/Interface/Internal/responseUtil'
-import { GetGamesRequest, GetGamesResponse } from '../../Common/Interface/Internal/games'
-import logger from '../logger'
-import { GetDbGamesByPlayerPuuid } from '../src/db/games'
-import { RiotMatchCallbackDto } from '../../Common/Interface/RiotAPI/RiotApiDto'
-import { SaveSingleMatchById } from '../src/business/games'
-import { ToMatchId } from '../../Common/utils'
+import { TypedRequest, TypedResponse } from '../../../Common/Interface/Internal/responseUtil'
+import { GetGamesRequest, GetGamesResponse } from '../../../Common/Interface/Internal/games'
+import logger from '../../logger'
+import { GetDbGamesByPlayerPuuid } from '../db/games'
+import { RiotMatchCallbackDto } from '../../../Common/Interface/RiotAPI/RiotApiDto'
+import { SaveSingleMatchById } from '../business/games'
+import { ToMatchId } from '../../../Common/utils'
 
 const router: Router = express.Router()
