+++ BackEnd/src/api/player.ts
@@ -1,14 +1,14 @@
 import express, { Request, Router } from "express";
-import { TypedRequest, TypedResponse } from "../../Common/Interface/Internal/responseUtil";
-import { PlayerChampionStatsResponse, PlayerGamesResponse, PlayerOverviewResponse, UpdatePlayerGamesResponse } from "../../Common/Interface/Internal/player";
-import { CreateChampionStatDataByPuuid, GetOrCreatePlayerOverviewByName, GetPlayerDetailedGames, UpdateGamesByPlayerPuuid } from "../src/business/player";
-import logger from "../logger";
-import { DocumentNotFound } from "../../Common/errors";
-import { GetDbChampionStatsByPlayerPuuid } from "../src/db/player";
-import { NonNone } from "../../Common/utils";
-import { GetGamesRequest } from "../../Common/Interface/Internal/games";
-import { GameRoles } from "../../Common/Interface/General/gameEnums";
-import {CreatePlayerStatsByPuuid} from "../src/business/playerstats";
+import { TypedRequest, TypedResponse } from "../../../Common/Interface/Internal/responseUtil";
+import { PlayerChampionStatsResponse, PlayerGamesResponse, PlayerOverviewResponse, UpdatePlayerGamesResponse } from "../../../Common/Interface/Internal/player";
+import { CreateChampionStatDataByPuuid, GetOrCreatePlayerOverviewByName, GetPlayerDetailedGames, UpdateGamesByPlayerPuuid } from "../business/player";
+import logger from "../../logger";
+import { DocumentNotFound } from "../../../Common/errors";
+import { GetDbChampionStatsByPlayerPuuid } from "../db/player";
+import { NonNone } from "../../../Common/utils";
+import { GetGamesRequest } from "../../../Common/Interface/Internal/games";
+import { GameRoles } from "../../../Common/Interface/General/gameEnums";
+import {CreatePlayerStatsByPuuid} from "../business/playerstats";
 
 const router: Router = express.Router()
